Revert "kartaAddHomeFrida"
